Get rid of TerminalAccessible.
* TODO: Deleted useless file. There was nothing interesting.
* libvte/libvtemm.h: Removed terminalaccessible.h inclusion.
* libvte/libvtemm/.gitignore: Removed terminalaccessible from
ignored files.
* libvte/libvtemm/Makefile.am: Use dist_noinst_HEADERS instead of
EXTRA_DIST.
* libvte/libvtemm/private/.gitignore: Ignore specific files, not
all ending with `_p.h'.
* libvte/src/filelist.am: Removed terminalaccesible.hg from
files_hg, so it won't be built nor distributed. For now it will
remain only in git (like reaper) - it is not clear to me, if this
class should be wrapped and how it should be wrapped. I was
wondering if this shouldn't be internal in vte. I didn't see any
application using it, but if someone will need it wrapped, then
patches are welcome.